feat: Tratamento de Exceptions delete

+++ src/main/java/com/cauabarros/aulajpacurso/services/UserService.java
@@ -2,8 +2,11 @@ package com.cauabarros.aulajpacurso.services;
 
 import com.cauabarros.aulajpacurso.entities.User;
 import com.cauabarros.aulajpacurso.repositories.UserRepository;
+import com.cauabarros.aulajpacurso.services.exceptions.DatabaseException;
 import com.cauabarros.aulajpacurso.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
@@ -23,8 +26,18 @@ public class UserService {
     public User insert(User obj){
         return repository.save(obj);
     }
-    public void delete(Long id){
-        repository.deleteById(id);
+    public void delete(Long id) {
+        try {
+            if (!repository.existsById(id)) {
+                System.out.println("Erro: Usuário não encontrado!");
+                throw new ResourceNotFoundException(id);
+            }
+            repository.deleteById(id);
+        } catch (EmptyResultDataAccessException e) {
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
     public User update(Long id, User obj){
         User entity = repository.getReferenceById(id);
